Add correct commit hash for shadow code
The hash that we listed before is not available in the public Shadow repo. It must have been a private commit that got rebased at some point and so the commit hash changed. Based on the log file of the privcount model experiment, I believe the new hash contains the changes we used in our experiments.